refactor(web): flip infra → services imports (#535 follow-up 2)
The four infra modules that reached into services to find shared
helpers now depend on neutral infra-tier modules instead, so the
api → services → infra arrow points the right way:

- infra/tunnels/tunnel-client.ts: emit moves to infra/events/status-event-bus.ts
- infra/browser-host/host-state.ts: getBrowser replaced by infra/browser-host/browser-lookup.ts
  (dependency-inversion registry that browser-service populates)
  + browser-manager.ts shim deleted
- infra/agents/agent-pool.ts: bandHome/loadSettings/getAgentDefinition
  pulled from infra/db/queries/settings.ts via SettingsQueries + resolveAgentDefinition
- infra/lsp/lsp-manager.ts: resolveWorkspace replaced by WorkspaceQueries.findIdentity

services/watcher.ts is now a façade over the infra event bus that adds
the on-connect snapshot logic; emit and StatusEvent are re-exported so
existing services callers don't need to migrate imports.

Refs: https://github.com/band-app/band/issues/535
