+++ apps/web/src/server/infra/tunnels/tunnel-client.ts
@@ -1,7 +1,7 @@
 import { type ChildProcess, spawn } from "node:child_process";
 import { createLogger } from "@band-app/logger";
-import { emit } from "../../services/watcher";
 import { getToken } from "../auth/auth-token";
+import { emit } from "../events/status-event-bus";
 import { shellPath } from "../process/path";
 
 const log = createLogger("tunnel");
