+++ apps/web/src/server/services/watcher.ts
@@ -1,78 +1,38 @@
 import { getDb } from "../infra/db/connection";
 import { branchStatuses as branchStatusesTable } from "../infra/db/schema";
-import { getRunningSetups } from "../services/setup-runner";
+import {
+  emit,
+  listenerCount,
+  type StatusEvent,
+  type StatusListener,
+  subscribe as subscribeRaw,
+} from "../infra/events/status-event-bus";
 import { startBranchStatusPoller, stopBranchStatusPoller } from "./branch-status-poller";
-import { loadCurrentStatuses, type WorkspaceStatus } from "./state";
+import { getRunningSetups } from "./setup-runner";
+import { loadCurrentStatuses } from "./state";
 
-interface GitStatus {
-  dirty: boolean;
-  conflict: boolean;
-  ahead: number;
-  behind: number;
-  sync_state: string;
-}
+/**
+ * Services-tier façade over the infra status event bus
+ * (`infra/events/status-event-bus.ts`).
+ *
+ * The bare `emit` / `subscribe` primitives + the `StatusEvent` shape live
+ * in infra so the lower-level adapters (tunnel-client) can publish
+ * without crossing into the services tier. This module adds the parts
+ * that *do* belong in services:
+ *
+ *   - the on-connect snapshot replay (current workspace statuses, branch
+ *     statuses, running setups) — needs DB access + the setup-runner
+ *     registry, both services-tier concerns;
+ *   - the branch-status poller lifecycle (start on first subscribe, stop
+ *     when the last subscriber disconnects).
+ */
 
-interface CIStatus {
-  state: string;
-  url?: string | null;
-}
+export type { StatusEvent };
 
-export interface StatusEvent {
-  kind:
-    | "update"
-    | "remove"
-    | "snapshot"
-    | "branch-status"
-    | "tunnel-url"
-    | "tunnel-error"
-    | "setup-status"
-    | "browser-created"
-    | "browser-removed"
-    | "terminal-created"
-    | "terminal-killed"
-    | "chat-created"
-    | "chat-removed"
-    | "open-file";
-  status?: WorkspaceStatus;
-  statuses?: WorkspaceStatus[];
-  workspaceId?: string;
-  git?: GitStatus;
-  ci?: CIStatus;
-  url?: string;
-  error?: string;
-  setupState?: "running" | "completed" | "failed";
-  setupError?: string;
-  runningSetups?: string[];
-  browserId?: string;
-  terminalId?: string;
-  chatId?: string;
-  /**
-   * For `kind: "open-file"`: workspace-relative file path with optional
-   * line / column suffix in the standard `path:line[:column]` /
-   * `path:line-lineEnd` notation. Parsed by the client via
-   * `parseFileLocation` from `@/dashboard`. Backs the
-   * `band open` CLI command — see `editorRouter.openFile`.
-   */
-  filePath?: string;
-  /**
-   * For `kind: "open-file"`: whether to bring the dashboard window to
-   * the foreground in addition to navigating to the file. Defaults to
-   * true. Wired through the desktop IPC bridge by the renderer; the
-   * plain web build ignores the field.
-   */
-  focus?: boolean;
-  /**
-   * For `kind: "open-file"`: whether the file lives outside the
-   * resolved workspace's root. When true, `filePath` carries an
-   * absolute filesystem path and the renderer should open it as an
-   * external tab (same surface as desktop Cmd+O / "Open File…").
-   */
-  external?: boolean;
-}
-
-type StatusListener = (event: StatusEvent) => void;
-
-const listeners: Set<StatusListener> = new Set();
+// Re-export the raw `emit` so existing callers (`services/*`, the
+// `api/*` routers) don't need to migrate their import paths in lockstep
+// with this refactor.
+export { emit };
 
 function loadCurrentBranchStatuses(): StatusEvent[] {
   const db = getDb();
@@ -94,14 +54,13 @@ function loadCurrentBranchStatuses(): StatusEvent[] {
   }));
 }
 
-export function emit(event: StatusEvent) {
-  for (const listener of listeners) {
-    listener(event);
-  }
-}
-
+/**
+ * Subscribe to status events with the on-connect snapshot replay.
+ * Starts the branch-status poller on the first subscribe and stops it
+ * when the last subscriber disconnects.
+ */
 export function subscribe(listener: StatusListener): () => void {
-  listeners.add(listener);
+  const unsubscribeRaw = subscribeRaw(listener);
   startBranchStatusPoller();
 
   // Send current agent status snapshot (always include runningSetups for reconciliation)
@@ -120,8 +79,8 @@ export function subscribe(listener: StatusListener): () => void {
   }
 
   return () => {
-    listeners.delete(listener);
-    if (listeners.size === 0) {
+    unsubscribeRaw();
+    if (listenerCount() === 0) {
       stopBranchStatusPoller();
     }
   };
